Moved CAPI_GLE into LibOVR.lb
Grasp uses just a small part of LibOVRKernel and was directly compiling
the 3-4 source files into the local projects. I instead include now
those files into the LibOVR library. This makes it easier to maintain
consistancy between the LibOVR and LibOVRKernel files if and when we
change versions. Note that the 0.8 and 1.3 versions have not yet been
updated in this way and are therefore probably broken. But I don't see
any need to go back.